request barrier (testcafe part of the hammerhead #498 issue)
