+++ src/client/driver/index.js
@@ -19,7 +19,7 @@ import DriverStatus from './status';
 
 var transport         = hammerhead.transport;
 var Promise           = hammerhead.Promise;
-var XhrBarrier        = testCafeCore.XhrBarrier;
+var RequestBarrier    = testCafeCore.RequestBarrier;
 var pageUnloadBarrier = testCafeCore.pageUnloadBarrier;
 var eventUtils        = testCafeCore.eventUtils;
 var modalBackground   = testCafeUI.modalBackground;
@@ -43,7 +43,7 @@ export default class ClientDriver {
         this.contextStorage             = new ContextStorage(window, testRunId);
         this.beforeUnloadRaised         = false;
 
-        this.pageInitialXhrBarrier = new XhrBarrier();
+        this.pageInitialRequestBarrier = new RequestBarrier();
 
         pageUnloadBarrier.init();
         preventRealEvents();
@@ -82,7 +82,7 @@ export default class ClientDriver {
 
         eventUtils
             .documentReady()
-            .then(() => this.pageInitialXhrBarrier.wait(true))
+            .then(() => this.pageInitialRequestBarrier.wait(true))
             .then(() => {
                 var inCommandExecution = this.contextStorage.getItem(COMMAND_EXECUTING_FLAG);
